src/third_party/blink/web_tests/http/tests/priority-hints/script-module-initial-load.html

<script src=resources/common.js></script>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>

<script id=script1 type=module importance=high src=../resources/dummy.js></script>
<script id=script2 type=module importance=low src=../resources/dummy.js?1></script>
<script id=script3 type=module importance=auto src=../resources/dummy.js?2></script>
<script id=script4 type=module importance=xyz src=../resources/dummy.js?3></script>
<script id=script5 type=module src=../resources/dummy.js?4></script>

<script>
  async_test(t => {
    addEventListener('DOMContentLoaded', t.step_func(() => {
      const msg = 'all scripts were fetched by the preload scanner';
      assert_true(internals.isPreloaded(script1.src), msg);
      assert_true(internals.isPreloaded(script2.src), msg);
      assert_true(internals.isPreloaded(script3.src), msg);
      assert_true(internals.isPreloaded(script4.src), msg);
      assert_true(internals.isPreloaded(script5.src), msg);

      assert_equals(getPriority(script1.src), kHigh,
                    'high importance on module <script> has no effect');
      assert_equals(getPriority(script2.src), kLow,
                    'low importance on module <script> lowers the priority to kLow');
      assert_equals(getPriority(script3.src), kHigh,
                    'auto importance on module <script> has no effect');
      assert_equals(getPriority(script4.src), kHigh,
                    'invalid importance on module <script> has no effect');
      assert_equals(getPriority(script5.src), kHigh,
                    'missing importance on module <script> has no effect');

      t.done();
    }));
  }, 'importance attribute on module <script>s fetched by the ' +
     'preload scanner takes the correct effect on resource load priority');
</script>
